add barcodes_list_view.dart and barcodes_panel.dart
