+++ lib/barcodes_page.dart
@@ -1,7 +1,9 @@
+import 'package:barcode_widget/barcode_widget.dart';
 import 'package:flutter/material.dart';
 import 'barcode_item.dart';
-import 'package:barcode_widget/barcode_widget.dart';
 import 'barcode_detail_page.dart';
+import 'barcodes_list_view.dart';
+import 'barcodes_panel.dart';
 
 class BarcodesPage extends StatefulWidget {
   const BarcodesPage({
@@ -16,27 +18,16 @@ class BarcodesPage extends StatefulWidget {
 }
 
 class BarcodesPageState extends State<BarcodesPage> {
-  final TextEditingController _descriptionController = TextEditingController();
-  final TextEditingController _dataController = TextEditingController();
-
-  void _addBarcode() {
-    final description = _descriptionController.text;
-    final data = _dataController.text;
-
-    if (description.isNotEmpty && data.isNotEmpty) {
-      setState(() {
-        widget.barcodes.add(
-          BarcodeItem(
-            description: description,
-            data: data,
-            type: Barcode.ean13(),
-          ),
-        );
-      });
-
-      _descriptionController.clear();
-      _dataController.clear();
-    }
+  void _addBarcode(String description, String data) {
+    setState(() {
+      widget.barcodes.add(
+        BarcodeItem(
+          description: description,
+          data: data,
+          type: Barcode.ean13(),
+        ),
+      );
+    });
   }
 
   @override
@@ -98,75 +89,10 @@ class BarcodesPageState extends State<BarcodesPage> {
                 },
               ),
             ),
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Row(
-                children: [
-                  Expanded(
-                    child: TextField(
-                      controller: _descriptionController,
-                      decoration: const InputDecoration(
-                        labelText: 'Opis',
-                      ),
-                    ),
-                  ),
-                  const SizedBox(width: 10),
-                  Expanded(
-                    child: TextField(
-                      controller: _dataController,
-                      decoration: const InputDecoration(
-                        labelText: 'Kod kreskowy',
-                      ),
-                    ),
-                  ),
-                  const SizedBox(width: 10),
-                  ElevatedButton(
-                    onPressed: _addBarcode,
-                    child: const Text('Dodaj kod kreskowy'),
-                  ),
-                ],
-              ),
-            ),
+            AddBarcodeForm(onAddBarcode: _addBarcode),
           ],
         ),
       ),
-    );
-  }
-}
-
-class BarcodeItemWidget extends StatelessWidget {
-  const BarcodeItemWidget({
-    super.key,
-    required this.barcode,
-  });
-
-  final BarcodeItem barcode;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          barcode.description,
-          style: const TextStyle(
-            fontSize: 17,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        const SizedBox(height: 6),
-        BarcodeWidget(
-          barcode: barcode.type,
-          data: barcode.data,
-          width: 200,
-          height: 100,
-          drawText: true,
-          style: const TextStyle(
-            fontSize: 16,
-            letterSpacing: 1.5,
-          ),
-        ),
-      ],
     );
   }
 }
